Fix #128: avoid Marshal-ing exn value via closure
This commit fixes an observed discrepancy between running a TeX filter
straight after parsing an Ott file, or after caching it and reading the
cache.

I believe the issue is that the code for the exception handler inside
the Marshal'd closure captures a value for the exception which is not
guaranteed to be preserved:

https://ocaml.org/manual/5.3/api/Marshal.html#:~:text=Values%20of%20extensible,work%20as%20expected
https://stackoverflow.com/a/54872172

One can observe this by debug-printing the tags associated with the
exception using the diff mentioned in [this comment](https://github.com/ott-lang/ott/issues/128#issuecomment-2851186496).